TextMenu class integration
